Module niti_hal::port

GPIO & Pin control.

This module contains a Pins struct which represents all pins of the board. The Pins struct is most easily constructed using the niti_hal::pins!() macro:

let dp = niti_hal::Peripherals::take().unwrap();
let pins = niti_hal::pins!(dp);

Additionally, the mode submodule contains all valid types for the MODE generic parameter of a pin. The Pin type-alias represents a pin which can represent any of the pins dynamically (while usually each pin has its own type).

Check the documentation for avr_hal_generic::port::Pin for a detailed explanation of GPIO pins in avr-hal.

Modules

mode

GPIO pin modes

Structs

Pins

Pins of the Arduino Mega 2560 and Arduino Mega 1280.

Traits

PinMode

PinOps

Type Aliases

A0

A0

A1

A1

A2

A2

A3

A3

A4

A4

A5

A5

A6

A6

A7

A7

A8

A8

A9

A9

A10

A10

A11

A11

A12

A12

A13

A13

A14

A14

A15

A15

D0

D0 / RX0

D1

D1 / TX0

D2

D2

D3

D3

D4

D4

D5

D5

D6

D6

D7

D7

D8

D8

D9

D9

D10

D10

D11

D11

D12

D12

D13

D13

D14

D14 / TX3

D15

D15 / RX3

D16

D16 / TX2

D17

D17 / RX2

D18

D18 / TX1

D19

D19 / RX1

D20

D20 / SDA

D21

D21 / SCL

D22

D22

D23

D23

D24

D24

D25

D25

D26

D26

D27

D27

D28

D28

D29

D29

D30

D30

D31

D31

D32

D32

D33

D33

D34

D34

D35

D35

D36

D36

D37

D37

D38

D38

D39

D39

D40

D40

D41

D41

D42

D42

D43

D43

D44

D44

D45

D45

D46

D46

D47

D47

D48

D48

D49

D49

D50

D50

D51

D51

D52

D52

D53

D53

Pin

Type-alias for a pin type which can represent any concrete pin.